api_yamdb/requirements.txt
attrs==21.4.0
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
flake8==4.0.1
idna==3.3
iniconfig==1.1.1
isort==5.10.1
mccabe==0.6.1
mypy==0.942
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==0.13.1
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.1.0
pyparsing==3.0.8
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.1
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
